refactor: 함수 파일별로 나누기
