+++ docs/README.md
@@ -24,6 +24,6 @@
 - [x] 2개의 숫자에 대해 뺄셈이 가능하다.
 - [x] 2개의 숫자에 대해 곱셈이 가능하다.
 - [x] 2개의 숫자에 대해 나눗셈이 가능하다.
-- [ ] AC(All Clear)버튼을 누르면 0으로 초기화 한다.
+- [x] AC(All Clear)버튼을 누르면 0으로 초기화 한다.
 - [x] 숫자는 한번에 최대 3자리 수까지 입력 가능하다.
 - [x] 계산 결과를 표현할 때 소수점 이하는 버림한다.
