+++ src/js/index.ts
@@ -1,103 +1,6 @@
-const calculateOperator = (
-  total: number,
-  operator: string,
-  num: number
-): number => {
-  let result: number = 0;
+import { setCountOperationdataSet } from './operation-data-set.js';
 
-  if (operator === '+') result = total + num;
-  if (operator === '/') result = total / num;
-  if (operator === '%') result = total % num;
-  if (operator === '-') result = total - num;
-  if (operator === 'X') result = total * num;
-  return result;
-};
-
-const checkValidInput = (total: string): string => {
-  const result: any = total.match(
-    new RegExp('(\\-?[\\d]{1,3})(X|\\-|\\+|\\/|\\=)?(\\-?[\\d]{1,3})?')
-  );
-  if (result[2] === undefined) {
-    if (result[1] === undefined) {
-      return '';
-    }
-    return result[1];
-  }
-  if (result[3] === undefined) {
-    return result[1] + result[2];
-  }
-  return result[1] + result[2] + result[3];
-};
-
-const calculateResult = (total: string): number => {
-  const result: any = total.match(
-    new RegExp('(\\-?[\\d]{1,3})(X|\\-|\\+|\\/|\\=)(\\-?[\\d]{1,3})')
-  );
-  console.log(result);
-  return calculateOperator(Number(result[1]), result[2], Number(result[3]));
-};
-
-const digitClickEvent = (e: Event): void => {
-  const eventTarget: HTMLElement = e.target as HTMLElement;
-  const totalTarget: HTMLElement | null = document.getElementById('total');
-
-  if (Number(totalTarget!.innerText) === 0) {
-    document.getElementById('total')!.innerText = eventTarget.innerText;
-  } else {
-    const result = checkValidInput(
-      document.getElementById('total')!.innerText + eventTarget.innerText
-    );
-    if (result === '') {
-      return;
-    }
-    document.getElementById('total')!.innerText = result;
-  }
-};
-
-const getCountOperationdataSet = (): string => {
-  const operations = document.getElementsByClassName(
-    'operations'
-  )[0]! as HTMLElement;
-  return operations.dataset['count'] ? operations.dataset['count'] : '';
-};
-
-const setCountOperationdataSet = (count: string): void => {
-  const operations = document.getElementsByClassName(
-    'operations'
-  )[0]! as HTMLElement;
-  operations.dataset['count'] = count;
-};
-
-const operatorEvent = (e: Event): void => {
-  const eventTarget: HTMLElement = e.target as HTMLElement;
-  const totalTarget: HTMLElement | null = document.getElementById('total');
-
-  if (eventTarget.innerText === '=') {
-    setCountOperationdataSet('0');
-    document.getElementById('total')!.innerText = String(
-      Math.floor(calculateResult(document.getElementById('total')!.innerText))
-    );
-    return;
-  }
-
-  if (Number(totalTarget!.innerText) === 0) {
-    return;
-  } else if (
-    !['/', '-', 'X', '+'].includes(
-      totalTarget!.innerText[totalTarget!.innerText.length - 1]
-    )
-  ) {
-    if (getCountOperationdataSet() === '1') {
-      return;
-    }
-    setCountOperationdataSet('1');
-    document.getElementById('total')!.innerText += eventTarget.innerText;
-  }
-};
-
-const ACEvent = (): void => {
-  document.getElementById('total')!.innerText = '0';
-};
+import { digitClickEvent, operatorEvent, ACEvent } from './controller.js';
 
 export default function App(): void {
   setCountOperationdataSet('0');
